MEETSAM: add index.php, passwd.txt, validate.js
passwd.txt in form of name:password:started

validate.js is functional

index.php the login works and will start the timer cookie.
i was working on The function that will modify the txt file to signify the quiz was started.
I will send the algorithm i was imagining when I was working on it